Extract withDashboard HOC into an own js module
Also change withDashboard to be a composable HOC.
